remove pages/
