+++ client/index.html
@@ -7,6 +7,8 @@
     body {
       margin: 0;
       font-family: 'open sans', arial, sans-serif;
+      background: #1b1b1b;
+      color: #fff;
     }
   </style>
 </head>
